Introduce function for file download
Since I got rid of the download/new directory logic, there's no point in
storing the downloaded file in a variable. Just write it directly to a
file. Normally I'd write/read the file chunk-wise, but that's not really
necessary with 4chan's strict file size limit.

Also fix the missing slashes in the log message.