Switch from eternal loop to limited attempts
This is one of my biggest gripes with the original version. The script
wouldn't complete unless a thread 404'd, which might be useful for
people who let this script run in the background to collect the newest
files of active threads, but not for me as I'm only interested in
already finished threads.
Of course, even if you want to download a thread only once, there can
be various connection errors. So the main download still runs in a loop,
only that this loop is limited to a finite number of retries, which can
be set via the respective command line option.

Additionally the regex search was moved to its own try-except block.
Mainly because it's easier and faster to catch a missing connection that
way.