Replace regex search with JSON parsing
I'll be honest. I don't like regex. Never did. Understanding it is
probably what makes the difference between smart people and people like
myself. However, I'm still convinced that this little list expansion is
far more readable and easier to understand than a giant regex pattern.

In addition to regex, the --use-names option also had to go. The
functionality is still there. I just don't see the point of requiring
yet another option, when the user already explicitly requested a custom
directory name. If a custom name is appended to the URL it will be used,
otherwise not. Simple as that.