Remove partially downloaded files on interrupt
The problem when downloading several files at once is that if the script
gets interrupted (or fails because of some other reason), you end up
with a workspace littered with incomplete files.
Remedy that by using the .part suffix on yet incomplete files and delete
any .part files if the ClientSession fails (for whatever reason).