Print thread-wise progress
In addition to printing the file-wise progress per thread, also print
the thread-wise progress. Looks a bit prettier during execution and
allows to better estimate the overall progress.

I'm not happy with this solution though. Throwing around a global dict
and random f-string assignments is just ugly. The overall file count is
also still missing. But it's good enough for today.