Add option to use original filenames
A nice option to have. To make the implementation easier, files (as
stored by each thread object) are now dictionaries. Currently they only
store the direct link and the output name, but this will change in the
future (very important: MD5 hashes).
The list comprehension to create the file list was also moved from
DownloadableThread.gather_files() to the class' initialization. The
method (now with a different name of course) is now only responsible to
download the API JSON response.