Handle aiohttp.ClientPayloadErrors
According to aiohttp's documentation these errors occur for

  * invalid compression
  * malformed chunked encoding
  * not enough data that satisfy Content-Length HTTP header

These errors are rare, but they can happen randomly, so they should be
covered by the retry-loop as well.

And since I'm at it, also change the help text to make it more clear,
that the retry count is per thread and not global.