Keep track of downloaded files with MD5 hashes
Add support for archive files. Archive files are simply text files, used
to store MD5 hashes of downloaded files. Logging and checking is done
in DownloadableThread.get_file(), which means the archive file gets
opened at least once for each listed file. In theory we could reduce the
amount of reading by loading the archive's contents at the beginning of
each thread, but for now I keep it on a per-file basis.

A crude validity check is made in parse_cli(), by trying to read the
first byte of the file.