Introduce --quiet option
The purpose should be self-explanatory. With -q/--quiet all non-error
messages are suppressed. The verbosity system is based on verbosity
levels. If the verbosity level of a message is higher than the global
one then it won't be printed. Because of that it is also possible to
suppress all errors/warnings, by simply choosing a low enough default
verbosity level.